Move number formatter to a separate file
