+++ Sources/Charcoal/Filters/Range/RangeFilterValueFormatter.swift
@@ -3,20 +3,19 @@
 //
 
 final class RangeFilterValueFormatter: NSObject, SliderValueFormatter {
-    private let isValueCurrency: Bool
+    private let formatWithSeparator: Bool
     private let unit: String
     private let accessibilityUnit: String
 
-    init(isValueCurrency: Bool, unit: String, accessibilityUnit: String) {
-        self.isValueCurrency = isValueCurrency
+    init(formatWithSeparator: Bool, unit: String, accessibilityUnit: String) {
+        self.formatWithSeparator = formatWithSeparator
         self.unit = unit
         self.accessibilityUnit = accessibilityUnit
     }
 
     func string(from value: Int) -> String? {
-        if isValueCurrency {
-            let value = NSNumber(value: value)
-            return NumberFormatter.currencyFormatter.string(from: value)
+        if formatWithSeparator {
+            return NumberFormatter.formatterWithSeparator.string(from: value)
         } else {
             return "\(value)"
         }
@@ -29,17 +28,4 @@ final class RangeFilterValueFormatter: NSObject, SliderValueFormatter {
     func title<ValueKind>(for value: ValueKind) -> String {
         return "\(value) \(unit)"
     }
-}
-
-// MARK: - Private extensions
-
-extension NumberFormatter {
-    static let currencyFormatter: NumberFormatter = {
-        let formatter = NumberFormatter()
-        formatter.numberStyle = .currency
-        formatter.currencySymbol = ""
-        formatter.locale = Locale(identifier: "nb_NO")
-        formatter.maximumFractionDigits = 0
-        return formatter
-    }()
 }
