+++ Sources/Charcoal/Models/RangeFilterConfiguration.swift
@@ -20,7 +20,7 @@ public struct RangeFilterConfiguration: Equatable {
     public let accessibilityValueSuffix: String?
     public let usesSmallNumberInputFont: Bool
     public let displaysUnitInNumberInput: Bool
-    public let isCurrencyValueRange: Bool
+    public let formatWithSeparator: Bool
 
     let formatter: RangeFilterValueFormatter
     private let isIncremented: Bool
@@ -58,7 +58,7 @@ public struct RangeFilterConfiguration: Equatable {
                 accessibilityValueSuffix: String?,
                 usesSmallNumberInputFont: Bool,
                 displaysUnitInNumberInput: Bool,
-                isCurrencyValueRange: Bool) {
+                formatWithSeparator: Bool) {
         self.minimumValue = minimumValue
         self.maximumValue = maximumValue
         self.hasLowerBoundOffset = hasLowerBoundOffset
@@ -67,7 +67,7 @@ public struct RangeFilterConfiguration: Equatable {
         self.accessibilityValueSuffix = accessibilityValueSuffix
         self.usesSmallNumberInputFont = usesSmallNumberInputFont
         self.displaysUnitInNumberInput = displaysUnitInNumberInput
-        self.isCurrencyValueRange = isCurrencyValueRange
+        self.formatWithSeparator = formatWithSeparator
 
         switch valueKind {
         case let .incremented(increment):
@@ -82,7 +82,7 @@ public struct RangeFilterConfiguration: Equatable {
         }
 
         formatter = RangeFilterValueFormatter(
-            isValueCurrency: isCurrencyValueRange,
+            formatWithSeparator: formatWithSeparator,
             unit: unit,
             accessibilityUnit: accessibilityValueSuffix ?? ""
         )
